+++ tests/MatchupServiceTests/CreateRoomTests.cs
@@ -8,7 +8,7 @@ using servartur.Services;
 using Moq.EntityFrameworkCore;
 
 namespace servartur.Tests.MatchupServiceTests;
-public class MatchupServiceCreateRoomTests
+public class CreateRoomTests
 {
     private static DbContextOptions<GameDbContext> getDbOptions()
         => new DbContextOptionsBuilder<GameDbContext>()
