+++ tests/MatchupServiceTests/RemovePlayerTests.cs
@@ -8,7 +8,7 @@ using servartur.Services;
 using Moq.EntityFrameworkCore;
 
 namespace servartur.Tests.MatchupServiceTests;
-public class MatchupServiceRemovePlayerTests
+public class RemovePlayerTests
 {
     private static DbContextOptions<GameDbContext> getDbOptions()
         => new DbContextOptionsBuilder<GameDbContext>()
