+++ tests/MatchupServiceTests/StartGameTests.cs
@@ -11,7 +11,7 @@ using servartur.Exceptions;
 
 namespace servartur.Tests.MatchupServiceTests;
 
-public class MatchupServiceStartGameTests
+public class StartGameTests
 {
     private static DbContextOptions<GameDbContext> getDbOptions()
         => new DbContextOptionsBuilder<GameDbContext>()
